refactor: inline LangGraph example factories
