+++ external/langgraph/examples/calculator.py
@@ -9,6 +9,8 @@ import argparse
 import asyncio
 import json
 from pathlib import Path
+from typing import Any
+from typing import TypedDict
 
 from nemo_fabric import Fabric
 from nemo_fabric import FabricConfig
@@ -17,6 +19,49 @@ from nemo_fabric import MetadataConfig
 from nemo_fabric import RuntimeConfig
 from nemo_fabric import WorkflowConfig
 from nemo_fabric import WorkflowEntrypointConfig
+
+
+class CalculatorState(TypedDict):
+    """Per-invocation calculator state."""
+
+    user_id: str
+    operation: str
+    left: float
+    right: float
+    answer: Any
+
+
+async def build_graph(mcp_url: str) -> Any:
+    """Return an uncompiled graph that calls the configured calculator server."""
+
+    from langchain_mcp_adapters.client import MultiServerMCPClient
+    from langgraph.graph import END
+    from langgraph.graph import START
+    from langgraph.graph import StateGraph
+
+    client = MultiServerMCPClient(
+        {
+            "mcp_math": {
+                "transport": "streamable-http",
+                "url": mcp_url,
+            }
+        }
+    )
+
+    async def calculate(state: CalculatorState) -> dict[str, Any]:
+        tools = await client.get_tools(server_name="mcp_math")
+        tool_by_name = {tool.name: tool for tool in tools}
+        tool = tool_by_name.get(state["operation"])
+        if tool is None:
+            raise ValueError(f"Unsupported calculator operation {state['operation']!r}")
+        answer = await tool.ainvoke({"left": state["left"], "right": state["right"]})
+        return {"answer": answer}
+
+    graph = StateGraph(CalculatorState)
+    graph.add_node("calculate", calculate)
+    graph.add_edge(START, "calculate")
+    graph.add_edge("calculate", END)
+    return graph
 
 
 def build_config(mcp_url: str) -> FabricConfig:
@@ -34,7 +79,7 @@ def build_config(mcp_url: str) -> FabricConfig:
         workflow=WorkflowConfig(
             entrypoint=WorkflowEntrypointConfig(
                 kind="langgraph_factory",
-                ref="calculator_graph:build_graph",
+                ref="calculator:build_graph",
             ),
             settings={"mcp_url": mcp_url},
         ),
